System.OsString.Posix

Types

data PosixString

Commonly used Posix string as uninterpreted char[] array.

data PosixChar

String construction

encodeUtf :: MonadThrow m => String -> m PosixString

Partial unicode friendly encoding.

This encodes as UTF8 (strictly), which is a good guess.

Throws an EncodingException if encoding fails. If the input does not contain surrogate chars, you can use unsafeEncodeUtf.

unsafeEncodeUtf :: HasCallStack => String -> PosixString

Unsafe unicode friendly encoding.

Like encodeUtf, except it crashes when the input contains surrogate chars. For sanitized input, this can be useful.

encodeWith :: TextEncoding -> String -> Either EncodingException PosixString

Encode a String with the specified encoding.

encodeFS :: String -> IO PosixString

This mimics the behavior of the base library when doing filesystem operations, which uses shady PEP 383 style encoding (based on the current locale, but PEP 383 only works properly on UTF-8 encodings, so good luck).

Looking up the locale requires IO. If you're not worried about calls to setFileSystemEncoding, then unsafePerformIO may be feasible (make sure to deeply evaluate the result to catch exceptions).

fromBytes :: MonadThrow m => ByteString -> m PosixString

Constructs a platform string from a ByteString.

This is a no-op.

pstr :: QuasiQuoter

QuasiQuote a PosixString. This accepts Unicode characters and encodes as UTF-8 on unix.

singleton :: PosixChar -> PosixString

empty :: PosixString

pack :: [PosixChar] -> PosixString

Pack a list of platform words to a platform string.

Note that using this in conjunction with unsafeFromChar to convert from [Char] to platform string is probably not what you want, because it will truncate unicode code points.

String deconstruction

decodeUtf :: MonadThrow m => PosixString -> m String

Partial unicode friendly decoding.

This decodes as UTF8 (strictly), which is a good guess. Note that filenames on unix are encoding agnostic char arrays.

Throws a EncodingException if decoding fails.

decodeWith :: TextEncoding -> PosixString -> Either EncodingException String

Decode a PosixString with the specified encoding.

The String is forced into memory to catch all exceptions.

decodeFS :: PosixString -> IO String

This mimics the behavior of the base library when doing filesystem operations, which uses shady PEP 383 style encoding (based on the current locale, but PEP 383 only works properly on UTF-8 encodings, so good luck).

Looking up the locale requires IO. If you're not worried about calls to setFileSystemEncoding, then unsafePerformIO may be feasible (make sure to deeply evaluate the result to catch exceptions).

unpack :: PosixString -> [PosixChar]

Unpack a platform string to a list of platform words.

Word construction

unsafeFromChar :: Char -> PosixChar

Truncates to 1 octet.

Word deconstruction

toChar :: PosixChar -> Char

Converts back to a unicode codepoint (total).

Basic interface

snoc :: PosixString -> PosixChar -> PosixString

O(n) Append a byte to the end of a OsString

Since: 1.4.200.0

cons :: PosixChar -> PosixString -> PosixString

O(n) cons is analogous to (:) for lists.

Since: 1.4.200.0

last :: HasCallStack => PosixString -> PosixChar

O(1) Extract the last element of a OsString, which must be finite and non-empty. An exception will be thrown in the case of an empty OsString.

This is a partial function, consider using unsnoc instead.

Since: 1.4.200.0

tail :: HasCallStack => PosixString -> PosixString

O(n) Extract the elements after the head of a OsString, which must be non-empty. An exception will be thrown in the case of an empty OsString.

This is a partial function, consider using uncons instead.

Since: 1.4.200.0

uncons :: PosixString -> Maybe (PosixChar, PosixString)

O(n) Extract the head and tail of a OsString, returning Nothing if it is empty.

Since: 1.4.200.0

head :: HasCallStack => PosixString -> PosixChar

O(1) Extract the first element of a OsString, which must be non-empty. An exception will be thrown in the case of an empty OsString.

This is a partial function, consider using uncons instead.

Since: 1.4.200.0

init :: HasCallStack => PosixString -> PosixString

O(n) Return all the elements of a OsString except the last one. An exception will be thrown in the case of an empty OsString.

This is a partial function, consider using unsnoc instead.

Since: 1.4.200.0

unsnoc :: PosixString -> Maybe (PosixString, PosixChar)

O(n) Extract the init and last of a OsString, returning Nothing if it is empty.

Since: 1.4.200.0

null :: PosixString -> Bool

O(1). The empty OsString.

Since: 1.4.200.0

length :: PosixString -> Int

O(1) The length of a OsString.

This returns the number of code units (Word8 on unix and Word16 on windows), not bytes.

>>> length "abc"
3

Since: 1.4.200.0

Transforming OsString

map :: (PosixChar -> PosixChar) -> PosixString -> PosixString

O(n) map f xs is the OsString obtained by applying f to each element of xs.

Since: 1.4.200.0

reverse :: PosixString -> PosixString

O(n) reverse xs efficiently returns the elements of xs in reverse order.

Since: 1.4.200.0

intercalate :: PosixString -> [PosixString] -> PosixString

O(n) The intercalate function takes a OsString and a list of OsStrings and concatenates the list after interspersing the first argument between each element of the list.

Since: 1.4.200.0

Reducing OsStrings (folds)

foldl :: forall a. (a -> PosixChar -> a) -> a -> PosixString -> a

foldl, applied to a binary operator, a starting value (typically the left-identity of the operator), and a OsString, reduces the OsString using the binary operator, from left to right.

Since: 1.4.200.0

foldl' :: forall a. (a -> PosixChar -> a) -> a -> PosixString -> a

foldl' is like foldl, but strict in the accumulator.

Since: 1.4.200.0

foldl1 :: (PosixChar -> PosixChar -> PosixChar) -> PosixString -> PosixChar

foldl1 is a variant of foldl that has no starting value argument, and thus must be applied to non-empty OsStrings. An exception will be thrown in the case of an empty OsString.

Since: 1.4.200.0

foldl1' :: (PosixChar -> PosixChar -> PosixChar) -> PosixString -> PosixChar

foldl1' is like foldl1, but strict in the accumulator. An exception will be thrown in the case of an empty OsString.

Since: 1.4.200.0

foldr :: forall a. (PosixChar -> a -> a) -> a -> PosixString -> a

foldr, applied to a binary operator, a starting value (typically the right-identity of the operator), and a OsString, reduces the OsString using the binary operator, from right to left.

Since: 1.4.200.0

foldr' :: forall a. (PosixChar -> a -> a) -> a -> PosixString -> a

foldr' is like foldr, but strict in the accumulator.

Since: 1.4.200.0

foldr1 :: (PosixChar -> PosixChar -> PosixChar) -> PosixString -> PosixChar

foldr1 is a variant of foldr that has no starting value argument, and thus must be applied to non-empty OsStrings An exception will be thrown in the case of an empty OsString.

Since: 1.4.200.0

foldr1' :: (PosixChar -> PosixChar -> PosixChar) -> PosixString -> PosixChar

foldr1' is a variant of foldr1, but is strict in the accumulator.

Since: 1.4.200.0

Special folds

all :: (PosixChar -> Bool) -> PosixString -> Bool

O(n) Applied to a predicate and a OsString, all determines if all elements of the OsString satisfy the predicate.

Since: 1.4.200.0

any :: (PosixChar -> Bool) -> PosixString -> Bool

O(n) Applied to a predicate and a OsString, any determines if any element of the OsString satisfies the predicate.

Since: 1.4.200.0

concat :: [PosixString] -> PosixString

Generating and unfolding OsStrings

replicate :: Int -> PosixChar -> PosixString

O(n) replicate n x is a OsString of length n with x the value of every element. The following holds:

replicate w c = unfoldr w (\u -> Just (u,u)) c

Since: 1.4.200.0

unfoldr :: forall a. (a -> Maybe (PosixChar, a)) -> a -> PosixString

O(n), where n is the length of the result. The unfoldr function is analogous to the List 'unfoldr'. unfoldr builds a OsString from a seed value. The function takes the element and returns Nothing if it is done producing the OsString or returns Just (a,b), in which case, a is the next byte in the string, and b is the seed value for further production.

This function is not efficient/safe. It will build a list of [Word8] and run the generator until it returns Nothing, otherwise recurse infinitely, then finally create a OsString.

If you know the maximum length, consider using unfoldrN.

Examples:

   unfoldr (\x -> if x <= 5 then Just (x, x + 1) else Nothing) 0
== pack [0, 1, 2, 3, 4, 5]

Since: 1.4.200.0

unfoldrN :: forall a. Int -> (a -> Maybe (PosixChar, a)) -> a -> (PosixString, Maybe a)

O(n) Like unfoldr, unfoldrN builds a OsString from a seed value. However, the length of the result is limited by the first argument to unfoldrN. This function is more efficient than unfoldr when the maximum length of the result is known.

The following equation relates unfoldrN and unfoldr:

fst (unfoldrN n f s) == take n (unfoldr f s)

Since: 1.4.200.0

Substrings

Breaking strings

take :: Int -> PosixString -> PosixString

O(n) take n, applied to a OsString xs, returns the prefix of xs of length n, or xs itself if n > length xs.

Since: 1.4.200.0

takeEnd :: Int -> PosixString -> PosixString

O(n) takeEnd n xs is equivalent to drop (length xs - n) xs. Takes n elements from end of bytestring.

>>> takeEnd 3 "abcdefg"
"efg"
>>> takeEnd 0 "abcdefg"
""
>>> takeEnd 4 "abc"
"abc"

Since: 1.4.200.0

takeWhileEnd :: (PosixChar -> Bool) -> PosixString -> PosixString

Returns the longest (possibly empty) suffix of elements satisfying the predicate.

takeWhileEnd p is equivalent to reverse . takeWhile p . reverse.

Since: 1.4.200.0

takeWhile :: (PosixChar -> Bool) -> PosixString -> PosixString

Similar to takeWhile, returns the longest (possibly empty) prefix of elements satisfying the predicate.

Since: 1.4.200.0

drop :: Int -> PosixString -> PosixString

O(n) drop n xs returns the suffix of xs after the first n elements, or empty if n > length xs.

Since: 1.4.200.0

dropEnd :: Int -> PosixString -> PosixString

O(n) dropEnd n xs is equivalent to take (length xs - n) xs. Drops n elements from end of bytestring.

>>> dropEnd 3 "abcdefg"
"abcd"
>>> dropEnd 0 "abcdefg"
"abcdefg"
>>> dropEnd 4 "abc"
""

Since: 1.4.200.0

dropWhileEnd :: (PosixChar -> Bool) -> PosixString -> PosixString

Similar to dropWhileEnd, drops the longest (possibly empty) suffix of elements satisfying the predicate and returns the remainder.

dropWhileEnd p is equivalent to reverse . dropWhile p . reverse.

Since: 1.4.200.0

dropWhile :: (PosixChar -> Bool) -> PosixString -> PosixString

Similar to dropWhile, drops the longest (possibly empty) prefix of elements satisfying the predicate and returns the remainder.

Since: 1.4.200.0

break :: (PosixChar -> Bool) -> PosixString -> (PosixString, PosixString)

Similar to break, returns the longest (possibly empty) prefix of elements which do not satisfy the predicate and the remainder of the string.

break p is equivalent to span (not . p) and to (takeWhile (not . p) &&& dropWhile (not . p)).

Since: 1.4.200.0

breakEnd :: (PosixChar -> Bool) -> PosixString -> (PosixString, PosixString)

Returns the longest (possibly empty) suffix of elements which do not satisfy the predicate and the remainder of the string.

breakEnd p is equivalent to spanEnd (not . p) and to (takeWhileEnd (not . p) &&& dropWhileEnd (not . p)).

Since: 1.4.200.0

span :: (PosixChar -> Bool) -> PosixString -> (PosixString, PosixString)

Similar to span, returns the longest (possibly empty) prefix of elements satisfying the predicate and the remainder of the string.

span p is equivalent to break (not . p) and to (takeWhile p &&& dropWhile p).

Since: 1.4.200.0

spanEnd :: (PosixChar -> Bool) -> PosixString -> (PosixString, PosixString)

Returns the longest (possibly empty) suffix of elements satisfying the predicate and the remainder of the string.

spanEnd p is equivalent to breakEnd (not . p) and to (takeWhileEnd p &&& dropWhileEnd p).

We have

spanEnd (not . isSpace) "x y z" == ("x y ", "z")

and

spanEnd (not . isSpace) sbs
   ==
let (x, y) = span (not . isSpace) (reverse sbs) in (reverse y, reverse x)

Since: 1.4.200.0

splitAt :: Int -> PosixString -> (PosixString, PosixString)

O(n) splitAt n sbs is equivalent to (take n sbs, drop n sbs).

Since: 1.4.200.0

split :: PosixChar -> PosixString -> [PosixString]

O(n) Break a OsString into pieces separated by the byte argument, consuming the delimiter. I.e.

split 10  "a\nb\nd\ne" == ["a","b","d","e"]   -- fromEnum '\n' == 10
split 97  "aXaXaXa"    == ["","X","X","X",""] -- fromEnum 'a' == 97
split 120 "x"          == ["",""]             -- fromEnum 'x' == 120
split undefined ""     == []                  -- and not [""]

and

intercalate [c] . split c == id
split == splitWith . (==)

Since: 1.4.200.0

splitWith :: (PosixChar -> Bool) -> PosixString -> [PosixString]

O(n) Splits a OsString into components delimited by separators, where the predicate returns True for a separator element. The resulting components do not contain the separators. Two adjacent separators result in an empty component in the output. eg.

splitWith (==97) "aabbaca" == ["","","bb","c",""] -- fromEnum 'a' == 97
splitWith undefined ""     == []                  -- and not [""]

Since: 1.4.200.0

stripSuffix :: PosixString -> PosixString -> Maybe PosixString

O(n) The stripSuffix function takes two OsStrings and returns Just the remainder of the second iff the first is its suffix, and otherwise Nothing.

Since: 1.4.200.0

stripPrefix :: PosixString -> PosixString -> Maybe PosixString

O(n) The stripPrefix function takes two OsStrings and returns Just the remainder of the second iff the first is its prefix, and otherwise Nothing.

Since: 1.4.200.0

Predicates

isInfixOf :: PosixString -> PosixString -> Bool

Check whether one string is a substring of another.

Since: 1.4.200.0

isPrefixOf :: PosixString -> PosixString -> Bool

O(n) The isPrefixOf function takes two OsStrings and returns True

Since: 1.4.200.0

isSuffixOf :: PosixString -> PosixString -> Bool

O(n) The isSuffixOf function takes two OsStrings and returns True iff the first is a suffix of the second.

The following holds:

isSuffixOf x y == reverse x `isPrefixOf` reverse y

Since: 1.4.200.0

Search for arbitrary susbstrings

breakSubstring :: PosixString -> PosixString -> (PosixString, PosixString)

Break a string on a substring, returning a pair of the part of the string prior to the match, and the rest of the string.

The following relationships hold:

break (== c) l == breakSubstring (singleton c) l

For example, to tokenise a string, dropping delimiters:

tokenise x y = h : if null t then [] else tokenise x (drop (length x) t)
    where (h,t) = breakSubstring x y

To skip to the first occurrence of a string:

snd (breakSubstring x y)

To take the parts of a string before a delimiter:

fst (breakSubstring x y)

Note that calling `breakSubstring x` does some preprocessing work, so you should avoid unnecessarily duplicating breakSubstring calls with the same pattern.

Since: 1.4.200.0

Searching OsStrings

Searching by equality

elem :: PosixChar -> PosixString -> Bool

O(n) elem is the OsString membership predicate.

Since: 1.4.200.0

find :: (PosixChar -> Bool) -> PosixString -> Maybe PosixChar

O(n) The find function takes a predicate and a OsString, and returns the first element in matching the predicate, or Nothing if there is no such element.

find f p = case findIndex f p of Just n -> Just (p ! n) ; _ -> Nothing

Since: 1.4.200.0

filter :: (PosixChar -> Bool) -> PosixString -> PosixString

O(n) filter, applied to a predicate and a OsString, returns a OsString containing those characters that satisfy the predicate.

Since: 1.4.200.0

partition :: (PosixChar -> Bool) -> PosixString -> (PosixString, PosixString)

O(n) The partition function takes a predicate a OsString and returns the pair of OsStrings with elements which do and do not satisfy the predicate, respectively; i.e.,

partition p bs == (filter p sbs, filter (not . p) sbs)

Since: 1.4.200.0

Indexing OsStrings

index :: HasCallStack => PosixString -> Int -> PosixChar

O(1) OsString index (subscript) operator, starting from 0.

Since: 1.4.200.0

indexMaybe :: PosixString -> Int -> Maybe PosixChar

O(1) OsString index, starting from 0, that returns Just if:

0 <= n < length bs

Since: 1.4.200.0

(!?) :: PosixString -> Int -> Maybe PosixChar

O(1) OsString index, starting from 0, that returns Just if:

0 <= n < length bs

Since: 1.4.200.0

elemIndex :: PosixChar -> PosixString -> Maybe Int

O(n) The elemIndex function returns the index of the first element in the given OsString which is equal to the query element, or Nothing if there is no such element.

Since: 1.4.200.0

elemIndices :: PosixChar -> PosixString -> [Int]

O(n) The elemIndices function extends elemIndex, by returning the indices of all elements equal to the query element, in ascending order.

Since: 1.4.200.0

count :: PosixChar -> PosixString -> Int

count returns the number of times its argument appears in the OsString

Since: 1.4.200.0

findIndex :: (PosixChar -> Bool) -> PosixString -> Maybe Int

O(n) The findIndex function takes a predicate and a OsString and returns the index of the first element in the OsString satisfying the predicate.

Since: 1.4.200.0

findIndices :: (PosixChar -> Bool) -> PosixString -> [Int]

O(n) The findIndices function extends findIndex, by returning the indices of all elements satisfying the predicate, in ascending order.

Since: 1.4.200.0